requirements.txt
aiogram==3.22.0
gspread==6.2.1
openai==2.6.0
protobuf==6.33.0
python-dotenv==1.1.1
